pyproject.toml
[project]
name = "llm-review"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "llama-cpp-python>=0.3.20",
]
